auto-grade-entry/requirements.txt
pandas
openpyxl
python-dotenv
playwright
browser-use
